requirements.txt
fastapi
uvicorn
pydantic
httpx
openai
pandas
requests
python-dotenv
openpyxl
nltk
pymongo
python-jose
passlib
google-auth
pyjwt
python-multipart
bcrypt